src/chatbot_engine/nlp_backend_requirements.txt
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
networkx>=3.0
numpy>=1.24.0
chromadb>=0.4.0
pathlib
logging
dataclasses
enum34
json5
time
